node_ref::dfs uses the refactored iterator
