+++ src/traversal/mod.rs
@@ -1,17 +1,17 @@
 pub(crate) mod depth_first;
 mod enumeration;
 mod enumerations;
+mod factory;
 mod node_item;
 mod node_item_mut;
 mod over;
 mod over_mut;
 pub(crate) mod post_order;
-mod traversal_factory;
 mod traverser;
 mod traverser_mut;
 
 pub use enumeration::Enumeration;
 pub use enumerations::{DepthSiblingIdxVal, DepthVal, SiblingIdxVal, Val};
-pub use traversal_factory::Traversal;
+pub use factory::Traversal;
 pub use traverser::Traverser;
 pub use traverser_mut::TraverserMut;
